SDK 3.2RC1, removed deprecated API usage, cleaned up project structures, added additional clustering sample

+++ hellomap3/Nuti.framework/Versions/A/Headers/datasources/PersistentCacheTileDataSource.h
@@ -50,9 +50,9 @@ namespace Nuti {
 
     protected:
         struct CacheElement {
-            CacheElement(long long tileId, unsigned int sizeInBytes);
+            CacheElement(long long tileId, size_t sizeInBytes);
             long long _tileId;
-            unsigned int _sizeInBytes;
+            size_t _sizeInBytes;
         };
         
         typedef std::list<CacheElement> CacheElementList;
